升级1.0.3

+++ mt-tools/video-tools/src/main/java/mt/spring/tools/video/entity/VideoInfo.java
@@ -14,7 +14,7 @@ public class VideoInfo {
 	private int width;
 	private int height;
 	private float frameRate;
-	private int bitRate;
+	private long bitRate;
 	private String decoder;
 	
 }
